dash_app/requirements.txt
argcomplete==1.12.2
boto3==1.16.59
botocore==1.19.59
Brotli==1.0.9
certifi==2020.12.5
cfn-flip==1.2.3
chardet==4.0.0
click==7.1.2
dash==1.19.0
dash-core-components==1.15.0
dash-html-components==1.1.2
dash-renderer==1.9.0
dash-table==4.11.2
durationpy==0.5
Flask==1.1.2
Flask-Compress==1.8.0
future==0.18.2
hjson==3.0.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
kappa==0.6.0
MarkupSafe==1.1.1
numpy==1.19.5
pandas==1.2.1
pip-tools==5.5.0
placebo==0.9.0
plotly==4.14.3
psycopg2-binary==2.8.6
python-dateutil==2.6.1
python-slugify==4.0.1
pytz==2020.5
PyYAML==5.4.1
requests==2.25.1
retrying==1.3.3
s3transfer==0.3.4
six==1.15.0
text-unidecode==1.3
toml==0.10.2
tqdm==4.56.0
troposphere==2.6.3
urllib3==1.26.2
Werkzeug==0.16.1
wsgi-request-logger==0.4.6
zappa==0.50.0
